- Reduced code duplication by creating the NNCONV - Passing NNConv integration test on Ray
