+++ applications/graph/GNN/test/conftest.py
@@ -9,14 +9,11 @@ def pytest_addoption(parser):
         'hostname'.split()).decode('utf-8').strip())
     default_dirname = subprocess.check_output(
         'git rev-parse --show-toplevel'.split()).decode('utf-8').strip()
-    default_exes = tools.get_default_exes(default_dirname, cluster)
 
     parser.addoption('--cluster', action='store', default=cluster,
                      help='--cluster=<cluster> to specify the cluster being run on, for the purpose of determing which commands to use. Default the current cluster')
     parser.addoption('--dirname', action='store', default=default_dirname,
                      help='--dirname=<path_to_dir> to specify the top-level directory. Default directory of build_lbann_lc executable')
-    parser.addoption('--exes', action='store', default=default_exes,
-                     help='--exes={compiler_name: path}')
     parser.addoption('--weekly', action='store_true', default=False,
                      help='--weekly specifies that the test should ONLY be run weekly, not nightly. Default False')
 
@@ -29,12 +26,6 @@ def cluster(request):
 @pytest.fixture
 def dirname(request):
     return request.config.getoption('--dirname')
-
-
-@pytest.fixture
-def exes(request):
-    return request.config.getoption('--exes')
-
 
 @pytest.fixture
 def weekly(request):
